# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /problem-reports:
    post:
      operationId: createProblemReport
      tags: [reports]
      summary: Submit a problem report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProblemReportBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                required:
                  - id
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request an upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
...
  /admin/problem-reports/{id}/status:
    patch:
      operationId: updateProblemReportStatus
      tags: [admin]
      summary: Update problem report status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                adminNote:
                  type: string
              required:
                - status
      responses:
        "200":
          description: Updated
  /admin/problem-reports/{id}/reply:
    patch:
      operationId: replyToProblemReport
      tags: [admin]
      summary: Reply to a problem report (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                replyText:
                  type: string
              required:
                - replyText
      responses:
        "200":
          description: Updated
...
  /plants/verify:
    post:
      operationId: verifyPlant
      tags: [trees]
      summary: Verify plant in image via AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPlantBody"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantVerificationResult"
components:
  schemas:
...
    CreateProblemReportBody:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
      required:
        - category
        - description
...
    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType
    UploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string
          required:
            - name
            - size
            - contentType
      required:
        - uploadURL
        - metadata
...
    VerifyPlantBody:
      type: object
      properties:
        imageBase64:
          type: string
        mimeType:
          type: string
      required:
        - imageBase64
        - mimeType
    PlantVerificationResult:
      type: object
      properties:
        isPlant:
          type: boolean
          nullable: true
        label:
          type: string
        reason:
          type: string
        model:
          type: string
        aiUnavailable:
          type: boolean
      required:
        - label
        - reason